+++ projects/angular-sdk/src/components/cfar-offer-banner-large/cfar-offer-banner-large.component.ts
@@ -93,7 +93,7 @@ export class CfarOfferBannerLargeComponent extends GlobalEventComponent implemen
                 this.isLoading = false;
               },
               error: (error) => {
-                console.error(error);
+                this.pushSdkError(error, "contracts");
                 this.isLoading = false;
               }
             });
@@ -127,6 +127,7 @@ export class CfarOfferBannerLargeComponent extends GlobalEventComponent implemen
 
   protected initCfarOffers(): void {
     this.isLoading = true;
+    console.log(this.hCSessionId)
     this._hopperCfarService
       .postCfarOffers(this.basePath, this.hCSessionId, this.currentLang, ApiTranslatorUtils.modelToSnakeCase(this._buildCreateCfarOfferRequest(this.itineraries)))
       .pipe(take(1))
@@ -150,7 +151,7 @@ export class CfarOfferBannerLargeComponent extends GlobalEventComponent implemen
           this.createPurchaseEventsAfterInit(this.hasWarningCoverageMessage);
         },
         error: (error) => {
-          console.error(error);
+          this.pushSdkError(error, "offers");
           this.offersLoaded.emit();
           this.isLoading = false;
         }