+++ projects/angular-sdk/src/components/cfar-offer-banner/cfar-offer-banner.component.ts
@@ -70,7 +70,7 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
             this.isLoading = false;
           },
           error: (error: any) => {
-            console.error(error);
+            this.pushSdkError(error, "offers");
             this.offersLoaded.emit();
             this.isLoading = false;
           }